refactor(podman): group utils files
