+++ extensions/podman/packages/extension/src/utils/podman-configuration.ts
@@ -25,8 +25,8 @@ import * as extensionApi from '@podman-desktop/api';
 import { Mutex } from 'async-mutex';
 import * as toml from 'smol-toml';
 
-import type { RegistryConfiguration } from './configuration/registry-configuration';
-import { RegistryConfigurationImpl } from './configuration/registry-configuration';
+import type { RegistryConfiguration } from '../configuration/registry-configuration';
+import { RegistryConfigurationImpl } from '../configuration/registry-configuration';
 
 const configurationRosetta = 'setting.rosetta';
