+++ extensions/podman/packages/extension/src/utils/podman-machine-stream.spec.ts
@@ -19,7 +19,7 @@
 import { Client } from 'ssh2';
 import { beforeEach, describe, expect, test, vi } from 'vitest';
 
-import type { MachineInfo } from './extension';
+import type { MachineInfo } from '../extension';
 import { ProviderConnectionShellAccessImpl } from './podman-machine-stream';
 
 const onMock = vi.fn();
