+++ extensions/podman/packages/extension/src/utils/registry-setup.ts
